requirements.txt
streamlit
deepface
pandas
pillow
openpyxl
pytz
opencv-python-headless
keras<3.0
